backend/requirements.txt
Django==1.10.2
django-extensions==1.7.4
djangorestframework==3.5.1
ipython==5.1.0
Pillow==3.4.2
requests==2.11.1
